Refactors use of substring()
This commit addresses https://github.com/mojolly/scala-inflector/issues/1 and a
number of other problems.

In many places the existing implementation assumes that a string will be at least one
character long by using `s.substring(0, 1)` and `s.substring(1)`. The tests missed
this because they did not include edge case, e.g. `""` or `"_"`, checks.

This commit refactors the use of `substring` to `take` and `slice`, which do not
raise exceptions. It also refactors `pascalize()` for more explicit handling of
`"__"`, which generate `""` elements in the list in the current implementation.

This is my original work, which I license to the rl project under the MIT License.